split local lab crate modules
